+++ packages/expo-ui/android/src/main/java/expo/modules/ui/CarouselView.kt
@@ -2,7 +2,6 @@
 
 package expo.modules.ui
 
-import android.content.Context
 import androidx.compose.foundation.gestures.TargetedFlingBehavior
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -11,19 +10,16 @@ import androidx.compose.material3.carousel.HorizontalUncontainedCarousel
 import androidx.compose.material3.carousel.rememberCarouselState
 import androidx.compose.material3.carousel.CarouselDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import androidx.core.view.size
-import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.apifeatures.EitherType
 import expo.modules.kotlin.types.Enumerable
 import expo.modules.kotlin.records.Field
 import expo.modules.kotlin.records.Record
 import expo.modules.kotlin.types.Either
+import androidx.core.view.size
 import expo.modules.kotlin.views.ComposeProps
-import expo.modules.kotlin.views.ExpoComposeView
+import expo.modules.kotlin.views.ExpoViewComposableScope
 import expo.modules.kotlin.views.ComposableScope
 
 enum class CarouselVariant(val value: String) : Enumerable {
@@ -72,15 +68,15 @@ fun paddingValuesFromEither(either: Either<Float, PaddingValuesRecord>?): Paddin
 }
 
 data class CarouselProps(
-  val variant: MutableState<CarouselVariant?> = mutableStateOf(null),
-  val modifiers: MutableState<List<ExpoModifier>?> = mutableStateOf(null),
-  val itemSpacing: MutableState<Float?> = mutableStateOf(null),
-  val contentPadding: MutableState<Either<Float, PaddingValuesRecord>?> = mutableStateOf(null),
-  val minSmallItemWidth: MutableState<Float?> = mutableStateOf(null),
-  val maxSmallItemWidth: MutableState<Float?> = mutableStateOf(null),
-  val flingBehavior: MutableState<FlingBehaviorType?> = mutableStateOf(null),
-  val preferredItemWidth: MutableState<Float?> = mutableStateOf(null),
-  val itemWidth: MutableState<Float?> = mutableStateOf(null)
+  val variant: CarouselVariant? = null,
+  val modifiers: List<ModifierConfig>? = null,
+  val itemSpacing: Float? = null,
+  val contentPadding: Either<Float, PaddingValuesRecord>? = null,
+  val minSmallItemWidth: Float? = null,
+  val maxSmallItemWidth: Float? = null,
+  val flingBehavior: FlingBehaviorType? = null,
+  val preferredItemWidth: Float? = null,
+  val itemWidth: Float? = null
 ) : ComposeProps
 
 const val DEFAULT_MIN_SMALL_ITEM_WIDTH = 40f
@@ -88,64 +84,60 @@ const val DEFAULT_MAX_SMALL_ITEM_WIDTH = 56f
 const val DEFAULT_PREFERRED_ITEM_WIDTH = 200f
 const val DEFAULT_ITEM_WIDTH = 200f
 
-class CarouselView(context: Context, appContext: AppContext) : ExpoComposeView<CarouselProps>(context, appContext) {
-  override val props = CarouselProps()
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ExpoViewComposableScope.CarouselContent(props: CarouselProps) {
+  val variant = props.variant ?: CarouselVariant.MULTI_BROWSE
+  val modifiers = props.modifiers ?: emptyList()
+  val itemSpacing = (props.itemSpacing ?: 0f).dp
+  val minSmallItemWidth = (props.minSmallItemWidth ?: DEFAULT_MIN_SMALL_ITEM_WIDTH).dp
 
-  @OptIn(ExperimentalMaterial3Api::class)
+  // we need to constrain maxSmallItemWidth to be at least minSmallItemWidth or the app will crash
+  val maxSmallItemWidth = minSmallItemWidth.coerceAtLeast((props.maxSmallItemWidth ?: DEFAULT_MAX_SMALL_ITEM_WIDTH).dp)
+  val preferredItemWidth = (props.preferredItemWidth ?: DEFAULT_PREFERRED_ITEM_WIDTH).dp
+  val itemWidth = (props.itemWidth ?: DEFAULT_ITEM_WIDTH).dp
+  val flingBehaviorType = props.flingBehavior ?: FlingBehaviorType.SINGLE_ADVANCE
+  val contentPadding = paddingValuesFromEither(props.contentPadding)
+
+  val carouselState = rememberCarouselState(0) { view.size }
+
+  val flingBehavior: TargetedFlingBehavior = when (flingBehaviorType) {
+    FlingBehaviorType.SINGLE_ADVANCE -> CarouselDefaults.singleAdvanceFlingBehavior(state = carouselState)
+    FlingBehaviorType.NO_SNAP -> CarouselDefaults.noSnapFlingBehavior()
+  }
+
   @Composable
-  override fun ComposableScope.Content() {
-    val variant = props.variant.value ?: CarouselVariant.MULTI_BROWSE
-    val modifiers = props.modifiers.value ?: emptyList()
-    val itemSpacing = (props.itemSpacing.value ?: 0f).dp
-    val minSmallItemWidth = (props.minSmallItemWidth.value ?: DEFAULT_MIN_SMALL_ITEM_WIDTH).dp
-
-    // we need to constrain maxSmallItemWidth to be at least minSmallItemWidth or the app will crash
-    val maxSmallItemWidth = minSmallItemWidth.coerceAtLeast((props.maxSmallItemWidth.value ?: DEFAULT_MAX_SMALL_ITEM_WIDTH).dp)
-    val preferredItemWidth = (props.preferredItemWidth.value ?: DEFAULT_PREFERRED_ITEM_WIDTH).dp
-    val itemWidth = (props.itemWidth.value ?: DEFAULT_ITEM_WIDTH).dp
-    val flingBehaviorType = props.flingBehavior.value ?: FlingBehaviorType.SINGLE_ADVANCE
-    val contentPadding = paddingValuesFromEither(props.contentPadding.value)
-
-    val carouselState = rememberCarouselState(0) { size }
-
-    val flingBehavior: TargetedFlingBehavior = when (flingBehaviorType) {
-      FlingBehaviorType.SINGLE_ADVANCE -> CarouselDefaults.singleAdvanceFlingBehavior(state = carouselState)
-      FlingBehaviorType.NO_SNAP -> CarouselDefaults.noSnapFlingBehavior()
+  fun MultiBrowseCarouselComposable() {
+    HorizontalMultiBrowseCarousel(
+      state = carouselState,
+      preferredItemWidth = preferredItemWidth,
+      modifier = ModifierRegistry.applyModifiers(modifiers),
+      itemSpacing = itemSpacing,
+      flingBehavior = flingBehavior,
+      minSmallItemWidth = minSmallItemWidth,
+      maxSmallItemWidth = maxSmallItemWidth,
+      contentPadding = contentPadding
+    ) { itemIndex ->
+      Child(ComposableScope(), itemIndex)
     }
+  }
 
-    @Composable
-    fun MultiBrowseCarouselComposable() {
-      HorizontalMultiBrowseCarousel(
-        state = carouselState,
-        preferredItemWidth = preferredItemWidth,
-        modifier = Modifier.fromExpoModifiers(modifiers, this@Content),
-        itemSpacing = itemSpacing,
-        flingBehavior = flingBehavior,
-        minSmallItemWidth = minSmallItemWidth,
-        maxSmallItemWidth = maxSmallItemWidth,
-        contentPadding = contentPadding
-      ) { itemIndex ->
-        Child(ComposableScope(), itemIndex)
-      }
+  @Composable
+  fun UnconstrainedCarouselComposable() {
+    HorizontalUncontainedCarousel(
+      state = carouselState,
+      itemWidth = itemWidth,
+      modifier = ModifierRegistry.applyModifiers(modifiers),
+      itemSpacing = itemSpacing,
+      flingBehavior = flingBehavior,
+      contentPadding = contentPadding
+    ) { itemIndex ->
+      Child(ComposableScope(), itemIndex)
     }
+  }
 
-    @Composable
-    fun UnconstrainedCarouselComposable() {
-      HorizontalUncontainedCarousel(
-        state = carouselState,
-        itemWidth = itemWidth,
-        modifier = Modifier.fromExpoModifiers(modifiers, this@Content),
-        itemSpacing = itemSpacing,
-        flingBehavior = flingBehavior,
-        contentPadding = contentPadding
-      ) { itemIndex ->
-        Child(ComposableScope(), itemIndex)
-      }
-    }
-
-    when (variant) {
-      CarouselVariant.MULTI_BROWSE -> MultiBrowseCarouselComposable()
-      CarouselVariant.UNCONSTRAINED -> UnconstrainedCarouselComposable()
-    }
+  when (variant) {
+    CarouselVariant.MULTI_BROWSE -> MultiBrowseCarouselComposable()
+    CarouselVariant.UNCONSTRAINED -> UnconstrainedCarouselComposable()
   }
 }