+++ packages/expo-ui/src/jetpack-compose/ContextMenu/index.tsx
@@ -4,14 +4,14 @@ import { NativeSyntheticEvent, StyleProp, ViewStyle } from 'react-native';
 
 import { SubmenuProps } from './Submenu';
 import { MenuElement, transformChildrenToElementArray } from './utils';
-import { ExpoModifier } from '../../types';
+import { ModifierConfig } from '../../types';
 import { ButtonProps } from '../Button';
 import { PickerProps } from '../Picker';
 import { SwitchProps } from '../Switch';
 
 const MenuNativeView: React.ComponentType<NativeMenuProps> = requireNativeView(
   'ExpoUI',
-  'ContextMenu'
+  'ContextMenuView'
 );
 
 type SubmenuElement =
@@ -60,7 +60,7 @@ export type ContextMenuProps = {
   /**
    * Modifiers for the component.
    */
-  modifiers?: ExpoModifier[];
+  modifiers?: ModifierConfig[];
 };
 
 /**
@@ -131,8 +131,7 @@ function ContextMenu(props: ContextMenuProps) {
       onContextMenuButtonPressed={createEventHandler('onPress')}
       onContextMenuSwitchValueChanged={createEventHandler('onValueChange')}
       onContextMenuPickerOptionSelected={createEventHandler('onOptionSelected')}
-      // @ts-expect-error
-      modifiers={props.modifiers?.map((m) => m.__expo_shared_object_id__)}
+      modifiers={props.modifiers}
       {...props}>
       {activationElement}
     </MenuNativeView>